+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.6] 2022-05-28
+
+### Added
+
+### Changed
+
+- The computation of scene's boundaries was removed from **CBoundsUpdatingSystem**. 
+
 ## [0.6.5] 2022-05-28
 
 ### Added